Move logic to k8s data provider
